// src/error-module/types/customErrors.ts
export enum CUSTOM_ERROR {
  INVALID_CREDENTIALS = "INVALID_CREDENTIALS",
  INVALID_TOKEN = "INVALID_TOKEN",

  FORBIDDEN_PERMISSION = "FORBIDDEN_PERMISSION",

  TOKEN_EXPIRED = "TOKEN_EXPIRED",

  USER_NOT_FOUND = "USER_NOT_FOUND",

  USER_ALREADY_EXIST = "USER_ALREADY_EXIST",
  PHONE_ALREADY_EXIST = "PHONE_ALREADY_EXIST",

  DUPLICATE_REGISTER = "DUPLICATE_REGISTER",
}